pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "shipsy-pf-builder"
version = "0.1.0"
description = "Process Flow Builder — turn a client meeting transcript into a Skynet-styled 3-slide PPTX deck."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Shipsy Pre-Sales" }
]
dependencies = [
    "anthropic>=0.40.0",
    "python-pptx>=0.6.23",
    "jsonschema>=4.21.0",
    "click>=8.1.0",
    "ulid-py>=1.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
pf = "pf_builder.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["pf_builder*"]
